Move steps under `file`. Wire up part formatter for different pipelines.
