+++ airbyte-cdk/bulk/toolkits/load-object-storage/src/main/kotlin/io/airbyte/cdk/load/pipline/object_storage/ObjectLoaderPartFormatterStep.kt
@@ -5,35 +5,33 @@
 package io.airbyte.cdk.load.pipline.object_storage
 
 import io.airbyte.cdk.load.factory.object_storage.ObjectKey
+import io.airbyte.cdk.load.message.DestinationRecordRaw
 import io.airbyte.cdk.load.message.PartitionedQueue
 import io.airbyte.cdk.load.message.PipelineEvent
+import io.airbyte.cdk.load.message.StreamKey
 import io.airbyte.cdk.load.pipeline.LoadPipelineStep
 import io.airbyte.cdk.load.task.internal.LoadPipelineStepTask
 import io.airbyte.cdk.load.task.internal.LoadPipelineStepTaskFactory
 import io.airbyte.cdk.load.write.object_storage.ObjectLoader
-import io.micronaut.context.annotation.Requires
-import jakarta.inject.Named
-import jakarta.inject.Singleton
 
-@Singleton
-@Requires(bean = ObjectLoader::class)
 class ObjectLoaderPartFormatterStep(
-    private val objectLoader: ObjectLoader,
+    objectLoader: ObjectLoader,
     private val partFormatter: ObjectLoaderPartFormatter<*>,
-    @Named("objectLoaderPartQueue")
-    val outputQueue:
-        PartitionedQueue<PipelineEvent<ObjectKey, ObjectLoaderPartFormatter.FormattedPart>>,
-    val taskFactory: LoadPipelineStepTaskFactory
+    private val inputQueue: PartitionedQueue<PipelineEvent<StreamKey, DestinationRecordRaw>>,
+    private val outputQueue: PartitionedQueue<PipelineEvent<ObjectKey, ObjectLoaderPartFormatter.FormattedPart>>,
+    private val taskFactory: LoadPipelineStepTaskFactory
 ) : LoadPipelineStep {
     override val numWorkers: Int = objectLoader.numPartWorkers
 
     override fun taskForPartition(partition: Int): LoadPipelineStepTask<*, *, *, *, *> {
-        return taskFactory.createFirstStep(
+        return taskFactory.createIntermediateStep(
             partFormatter,
+            inputQueue,
             ObjectLoaderFormattedPartPartitioner(),
             outputQueue,
             partition,
-            numWorkers
+            numWorkers,
+            1,
         )
     }
 }
