+++ airbyte-cdk/bulk/toolkits/load-object-storage/src/main/kotlin/io/airbyte/cdk/load/pipline/object_storage/ObjectLoaderPipeline.kt
@@ -7,6 +7,9 @@ package io.airbyte.cdk.load.pipline.object_storage
 import io.airbyte.cdk.load.file.object_storage.RemoteObject
 import io.airbyte.cdk.load.message.WithStream
 import io.airbyte.cdk.load.pipeline.LoadPipeline
+import io.airbyte.cdk.load.pipline.object_storage.file.FileChunkStep
+import io.airbyte.cdk.load.pipline.object_storage.file.ForwardFileRecordStep
+import io.airbyte.cdk.load.pipline.object_storage.file.RouteEventStep
 import io.airbyte.cdk.load.write.object_storage.ObjectLoader
 import io.micronaut.context.annotation.Requires
 import jakarta.inject.Named
